requirements.txt
kivy>=2.2.0
plyer==2.1.0
datetime
requests==2.27.1
python-dotenv==0.19.2
kivymd>=1.1.1
pillow>=9.0.0 